pyproject.toml
[tool.poetry]
name = "hint-cli"
version = "0.8.0"
description = "Get help and examples for commands from the command line, without switching context or applications."
authors = ["Andrew Garner"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/agarthetiger/hint/"
repository = "https://github.com/agarthetiger/hint/"
documentation = "https://agarthetiger.github.io/hint/"
keywords = ["cli", "help"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.7",
    "Topic :: Utilities",
]
include = [
    "LICENSE",
]
packages = [
    { include = "hint_cli" },
    { include = "markdown" },
]

[tool.poetry.dependencies]
python = "^3.7"
gitpython = "^3.1.7"
pygments = "~=2.7"
paramiko = "^2.7.2"
click = "~=8.0"


[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-cov = "^3.0.0"

[tool.poetry.scripts]
hint = 'hint_cli.hint:cli_entrypoint'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
